+++ cpp/src/strings/replace/replace.cu
@@ -377,7 +377,7 @@ std::unique_ptr<column> replace_nulls(
                              std::move(offsets_column),
                              std::move(chars_column),
                              0,
-                             rmm::device_buffer{},
+                             rmm::device_buffer{0, stream, mr},
                              stream,
                              mr);
 }